old-tests/webapps/XMLHttpRequest/tests/submissions/Opera/open-method-case-sensitive.htm

<!DOCTYPE html>
<html>
  <head>
    <title>XMLHttpRequest: open() - case-sensitive methods test</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      function method(method) {
        test(function() {
          var client = new XMLHttpRequest()
          client.open(method, "resources/content.php", false)
          client.send(null)
          assert_equals(client.getResponseHeader("x-request-method"), method)
        }, document.title + " (" + method + ")")
      }
      method("XUNICORN")
      method("xUNIcorn")
      method("chiCKEN")
      method("PATCH")
      method("patCH")
      method("copy")
      method("COpy")
      method("inDEX")
      method("movE")
    </script>
  </body>
</html>
